+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/intellij/settings/XPathColorSettingsPage.kt
@@ -32,7 +32,7 @@ class XPathColorSettingsPage : ColorSettingsPage {
 
     override fun getDemoText(): String = XPathQueries.ColorSettingsDemo
 
-    override fun getAdditionalHighlightingTagToDescriptorMap(): Map<String, TextAttributesKey>? = null
+    override fun getAdditionalHighlightingTagToDescriptorMap(): Map<String, TextAttributesKey>? = XPathSyntaxHighlighterColors.ADDITIONAL_DESCRIPTORS
 
     override fun getAttributeDescriptors(): Array<AttributesDescriptor> = XPathSyntaxHighlighterColors.DESCRIPTORS