+++ CHANGELOG.md
@@ -34,6 +34,7 @@ IntelliJ Integration:
 
 1.  Display the first comment line when the comment is folded.
 1.  Support profiling MarkLogic queries.
+1.  Support configuring the namespace prefix colour.
 
 Inspections: